+++ ui/shared/components/buttons/ShowGeneModal.jsx
@@ -3,11 +3,24 @@ import PropTypes from 'prop-types'
 
 import Modal from '../modal/Modal'
 import GeneDetail from '../panel/genes/GeneDetail'
-import { ButtonLink } from '../StyledComponents'
+import { ColoredIcon, ButtonLink } from '../StyledComponents'
+import { PANEL_APP_CONFIDENCE_LEVEL_COLORS, PANEL_APP_CONFIDENCE_DESCRIPTION } from '../../utils/constants'
 
-const ShowGeneModal = ({ gene, modalId = 'gene', ...linkProps }) =>
+const ShowGeneModal = ({ pagene, gene, modalId = 'gene', ...linkProps }) =>
   <Modal
-    trigger={<ButtonLink {...linkProps}>{gene.geneSymbol}</ButtonLink>}
+    trigger={
+      <ButtonLink
+        {...linkProps}
+      >
+        {pagene?.confidenceLevel > 0 &&
+          <ColoredIcon
+            name="circle"
+            title={PANEL_APP_CONFIDENCE_DESCRIPTION[pagene.confidenceLevel]}
+            color={PANEL_APP_CONFIDENCE_LEVEL_COLORS[pagene.confidenceLevel]}
+          />
+        }
+        {gene.geneSymbol}
+      </ButtonLink>}
     title={gene.geneSymbol}
     modalName={`${modalId}-${gene.geneId}`}
     size="fullscreen"
@@ -16,6 +29,7 @@ const ShowGeneModal = ({ gene, modalId = 'gene', ...linkProps }) =>
   </Modal>
 
 ShowGeneModal.propTypes = {
+  pagene: PropTypes.object,
   gene: PropTypes.object,
   modalId: PropTypes.string,
 }